requirement.txt
asgiref==3.7.2
bcrypt==4.0.1
DateTime==5.2
Django==4.2.4
django-cors-headers==4.2.0
django-filter==23.3
djangorestframework==3.14.0
et-xmlfile==1.1.0
gunicorn==21.2.0
mysqlclient==2.2.0
numpy==1.26.0
openpyxl==3.1.2
packaging==23.1
pandas==2.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
zope.interface==6.0
